requirements.txt
adal==1.2.2
aniso8601==8.1.0
azure-functions==1.7.2
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
bottle==0.12.18
cachelib==0.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
configparser==5.0.0
cryptography==2.9.2
decorator==4.4.2
dnspython==1.16.0
ecdsa==0.14.1
email-validator==1.1.0
et-xmlfile==1.1.0
etelemetry==0.2.2
filelock==3.0.12
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-OAuthlib==0.9.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fpdf==1.7.2
greenlet==1.1.0
idna==2.9
importlib-metadata==4.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyboard==0.13.5
lxml==4.5.2
MarkupSafe==1.1.1
marshmallow==3.5.2
msal==1.5.1
networkx==2.5
nibabel==3.1.1
numpy==1.18.4
oauthlib==2.1.0
openpyxl==3.0.9
packaging==20.4
pandas==1.0.3
passlib==1.7.2
Pillow==7.2.0
prov==1.5.3
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
pycrc==0.9.2
pydot==1.4.1
pydotplus==2.0.2
PyJWT==1.7.1
PyMuPDF==1.17.7
pyodbc==4.0.30
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.1
rdflib==5.0.0
requests==2.23.0
rsa==4.7.2
simplejson==3.17.2
six==1.14.0
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.3.16
typing-extensions==3.10.0.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
xlrd==1.2.0
zipp==3.4.1
